Backend/requirements.txt
allure-behave==2.9.43
allure-python-commons==2.9.43
asgiref==3.4.1
attrs==21.2.0
bcrypt==3.2.0
beautifulsoup4==4.9.3
behave==1.2.6
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.1
click==8.0.1
cryptography==3.4.7
fastapi==0.66.0
h11==0.12.0
idna==3.2
mysql-connector-python==8.0.25
paramiko==2.7.2
parse==1.19.0
parse-type==0.5.2
pluggy==0.13.1
protobuf==3.17.3
pycparser==2.20
pydantic==1.8.2
PyNaCl==1.4.0
requests==2.26.0
selenium==3.141.0
six==1.16.0
soupsieve==2.2.1
starlette==0.14.2
typing-extensions==3.10.0.0
Unidecode==1.2.0
urllib3==1.26.6
uvicorn==0.14.0
